+++ tests/esa_tests.c
@@ -13,9 +13,9 @@ char *test_getEsa() {
   mu_assert(esa->str == s, "ESA does not point to original sequence");
   mu_assert(esa->n == n + 1, "ESA size not correct");
   mu_assert(esa->str[esa->sa[0]] == '$', "first ESA entry not $");
-  mu_assert(esa->sa[0] == n, "wrong SA index");
+  mu_assert(esa->sa[0] == (int64_t)n, "wrong SA index");
   mu_assert(esa->isa[esa->sa[0]] == 0, "isa incorrect");
-  mu_assert(esa->isa[esa->sa[n]] == n, "isa incorrect");
+  mu_assert(esa->isa[esa->sa[n]] == (int64_t)n, "isa incorrect");
   mu_assert(esa->lcp[0] == -1, "first LCP not -1");
   mu_assert(esa->lcp[esa->n] == -1, "last LCP not -1");
